refactor: update paths for task lifecycle
