+++ src/app/network/UploadFolderManager.ts
@@ -17,7 +17,7 @@ import { wait } from 'utils/timeUtils';
 import { ConnectionLostError } from './requests';
 import { FilesExceedsSizeLimitError } from 'app/drive/services/file.service/upload.errors';
 import { filterFilesByMaxSize } from 'app/store/slices/storage/fileUtils/filterFilesByMaxSize';
-import { createUploadFolderTaskLifecycle, UploadFolderTaskCallbacks } from './upload/uploadFolderTaskLifecycle';
+import { createUploadFolderTaskLifecycle, UploadFolderTaskCallbacks } from 'app/tasks/upload/uploadFolderTaskLifecycle';
 
 interface UploadFolderPayload {
   root: IRoot;
